debug

+++ precision_recall_opt.py
@@ -96,18 +96,18 @@ def taxon_recall(taxon):
         if row['prediction'] == taxon:
             total_count += 1
             reject_start = math.floor(row['max']*1/gap)/(1/gap)+gap
-            reject_zero_half = [0]*(reject_start*1/gap)
-            reject_incre = reject_zero_half.extend([1]*(alpha_num-reject_start*1/gap))
-            correct_one_half = [1]*(reject_start*1/gap)
-            correct_incre  = correct_one_half.extend([0]*(alpha_num-reject_start*1/gap))
+            reject_zero_half = [0]*(int(reject_start*1/gap))
+            reject_incre = reject_zero_half.extend([1]*(int(alpha_num-reject_start*1/gap)))
+            correct_one_half = [1]*(int(reject_start*1/gap))
+            correct_incre  = correct_one_half.extend([0]*(int(alpha_num-reject_start*1/gap)))
             correct_stat = [sum(x) for x in zip(correct_stat, correct_incre)]
             reject_stat = [sum(x) for x in zip(reject_stat, reject_incre)]
             probs.append(row['max'])
     for index, row in w_df.iterrows():
         total_count += 1
         reject_start = math.floor(row['max']*1/gap)/(1/gap)+gap
-        reject_zero_half = [0]*(reject_start*1/gap)
-        reject_incre = reject_zero_half.extend([1]*(alpha_num-reject_start*1/gap))
+        reject_zero_half = [0]*(int(reject_start*1/gap))
+        reject_incre = reject_zero_half.extend([1]*(int(alpha_num-reject_start*1/gap)))
         reject_stat = [sum(x) for x in zip(reject_stat, reject_incre)] 
 
     # compute thres_alpha
